Add dialog templates

+++ src/Presentation/Proxirae.Presentation.WPF/ViewModels/ProxyServers/ProxyServersViewModel.cs
@@ -4,15 +4,17 @@ using MvvmDialogs;
 using Proxirae.Application.Services.Proxies;
 using Proxirae.Application.Validators.ProxyRule;
 using Proxirae.Contracts.DTOs.Proxies;
-using Proxirae.Presentation.WPF.Facades.Dialog;
 using Proxirae.Presentation.WPF.Factories.ProxyChecker;
+using Proxirae.Presentation.WPF.Services.Dialog.Message;
+using Proxirae.Presentation.WPF.Services.Dialog.Modal;
 using System.Collections.ObjectModel;
 
 namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
 {
     public partial class ProxyServersViewModel : ObservableObject, IModalDialogViewModel
     {
-        private readonly DialogFacade _dialogFacade;
+        private readonly IModalDialogService _modalDialogService;
+        private readonly IMessageDialogService _messageDialogService;
         private readonly IProxyService _proxyService;
         private readonly IProxyCheckerViewModelFactory _proxyCheckerFactory;
         private readonly IProxyRuleValidator _proxyRuleValidator;
@@ -30,9 +32,14 @@ namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
         [NotifyCanExecuteChangedFor(nameof(ApplyCommand))]
         private bool _hasChanges;
 
-        public ProxyServersViewModel(DialogFacade dialogFacade, IProxyService proxyService, IProxyCheckerViewModelFactory proxyCheckerFactory, IProxyRuleValidator proxyRuleValidator)
+        public ProxyServersViewModel(IModalDialogService dialogFacade,
+                                     IMessageDialogService messageDialogService, 
+                                     IProxyService proxyService,
+                                     IProxyCheckerViewModelFactory proxyCheckerFactory,
+                                     IProxyRuleValidator proxyRuleValidator)
         {
-            _dialogFacade = dialogFacade;
+            _modalDialogService = dialogFacade;
+            _messageDialogService = messageDialogService;
             _proxyService = proxyService;
             _proxyCheckerFactory = proxyCheckerFactory;
             _proxyRuleValidator = proxyRuleValidator;
@@ -56,7 +63,7 @@ namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
         [RelayCommand]
         private async Task AddProxyServerAsync(CancellationToken cancellationToken)
         {
-            var viewModel = _dialogFacade.ShowDialog<AddProxyServerViewModel>(this);
+            var viewModel = _modalDialogService.ShowDialog<AddProxyServerViewModel>(this);
 
             if (viewModel.ProxyServer is not null)
             {
@@ -76,7 +83,7 @@ namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
 
             var viewModel = new EditProxyServerViewModel(proxyDetail);
 
-            _dialogFacade.ShowDialog(this, viewModel);
+            _modalDialogService.ShowDialog(this, viewModel);
 
             if (viewModel.ProxyServer is not null)
             {
@@ -95,9 +102,15 @@ namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
         [RelayCommand]
         private async Task RemoveProxyServerAsync(ProxyListDto proxyServer, CancellationToken cancellationToken)
         {
+            var affectedRules = await _proxyRuleValidator.GetAffectedRulesAsync(proxyServer.Id, cancellationToken);
+            if (affectedRules.Count > 0 && !_messageDialogService.ShowWarning(this, "ProxyDelete", "ProxyDeleteTitle"))
+            {
+                return;
+            }
+
             await _proxyService.DeleteAsync(proxyServer.Id, cancellationToken);
-            await _proxyRuleValidator.UnbindProxyFromRulesAsync(proxyServer.Id, cancellationToken); // TODO: Add Confirmation Dialog menu
-            
+            await _proxyRuleValidator.UnbindProxyFromRulesAsync(proxyServer.Id, cancellationToken); 
+
             ProxyServers.Remove(proxyServer);
         }
 
@@ -111,7 +124,7 @@ namespace Proxirae.Presentation.WPF.ViewModels.ProxyServers
             }
 
             var viewModel = _proxyCheckerFactory.Create(detail);
-            _dialogFacade.ShowDialog(this, viewModel);
+            _modalDialogService.ShowDialog(this, viewModel);
         }
 
         [RelayCommand]
